change logic process_command to accept game state info
allow actions to be non-loggable

+++ src/gametime.rs
@@ -1,1 +1,1 @@
-pub type GameTime = usize;
+pub type GameTime = u64;
